mvghb/pyproject.toml
[project]
name = "mvghb"
version = "0.1.0"
description = "MVGH-β Wave 1 governance harness — sign, anchor, verify."
requires-python = ">=3.11"
dependencies = [
    "cryptography>=42",
    "psycopg[binary]>=3.1",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "click>=8.1",
    "structlog>=24.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
mvghb = "mvghb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mvghb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
